+++ CHANGES.md
@@ -8,6 +8,7 @@
    * Added camera view controls
    * Fixed notification config panel
    * Reworked desktop notification content to contain more information.
+   * Desktop notification timeouts can now be configured
 
 * v1.1.1 (2020-12-03)
    * Improved docs